+++ classes/controller/paginate/test.php
@@ -1,0 +1,36 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Paginate Test resource
+ *
+ * @author		Micheal Morgan <micheal.morgan@standvertical.com>
+ * @copyright	(c) 2011-2012 Stand Vertical Inc
+ */
+class Controller_Paginate_Test extends REST_Controller 
+{
+	/**
+	 * GET test
+	 * 
+	 * @access	public
+	 * @return	void
+	 */
+	public function action_get()
+	{
+		$database = DB::select()->from('paginate_test');
+		
+		$paginate = Paginate::factory($database)
+			->columns(array('id', 'label'));
+		
+		$collection = REST_Collection::factory($paginate)->execute();
+		
+		foreach ($collection->paginate()->result() as $row)
+		{
+			$collection->add_row(array
+			(
+				'id'	=> $row['id'],
+				'label'	=> $row['label']
+			));
+		}		
+
+		$this->render($collection->render());
+	}	
+}