+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Initial release of Paginate
 - Support for Kohana Database, Kohana ORM and Dispatch
 - Shipped with separate wrapper for DataTables
+- All tests pass "OK (6 tests, 12 assertions)"